functions/python/requirements.txt
Flask
PyJWT
bcrypt
google-cloud-datastore
